Make no-jinja-when also detect list conditions
Fixes: #2959
