--- examples/playbooks/jinja2-when-failure.yml
@@ -1,11 +1,0 @@
----
-- hosts: all
-  tasks:
-    - name: Test when with jinja2
-      ansible.builtin.debug: msg=text
-      when: "{{ false }}"
-
-- hosts: all
-  roles:
-    - role: test
-      when: "{{ '1' = '1' }}"
