+++ examples/playbooks/rule-no-jinja-when-pass.yml
@@ -1,0 +1,17 @@
+---
+- name: Test fixture for no-jinja-when
+  hosts: all
+  tasks:
+    - name: Test when
+      ansible.builtin.debug:
+        msg: text
+      when: true
+    - name: Test when 2
+      ansible.builtin.debug:
+        msg: text2
+      when: 1 = 1
+    - name: Three
+      ansible.builtin.debug:
+        msg: text2
+      when:
+        - "false"
